fix enemy

+++ TankOnlineFU/Assets/Scripts/TankFirer.cs
@@ -11,7 +11,7 @@ namespace DefaultNamespace
         public Sprite spriteLeft;
         public Sprite spriteUp;
         public Sprite spriteDown;
-        public int speed;
+        public float speed;
         public int maxRange;
         public float delay;
         public float lastFire = 0f;
@@ -30,7 +30,6 @@ namespace DefaultNamespace
             {
                 return;
             }
-
             var bullet = Instantiate(bulletPrefab, b.InitialPosition, Quaternion.identity);
             var sr = bullet.GetComponent<SpriteRenderer>();
             var rigidBody2d = bullet.GetComponent<Rigidbody2D>();
@@ -62,7 +61,6 @@ namespace DefaultNamespace
                 default:
                     throw new ArgumentOutOfRangeException();
             }
-
             rigidBody2d.AddForce(force, ForceMode2D.Impulse);
             lastFire = Time.time;
         }
